requirements.txt
discord.py
python-dotenv
mercapi
asyncio
